move generate_metadata_config back to cm module
