+++ pyproject.toml
@@ -37,6 +37,9 @@ check_license = "python -c 'from scripts.checks import check_license; check_lice
 license = "python -c 'from scripts.checks import license; license()'"
 set_version = "python -c 'from scripts.set_version import main; main()'"
 
+[tool.poetry.scripts]
+generate_metadata_config = "scripts.ontology.generate_metadata_config:main"
+
 [tool.black]
 line_length = 120
